+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 Summary of release changes for Version 2 - CentOS-7
 
+### 2.4.1 - 2018-11-10
+
+- Adds feature to set system time zone via `SSH_TIMEZONE`.
+- Adds feature to enable password authentication.
+- Adds default of removing insecure public key when enabling password authentication.
+
 ### 2.4.0 - 2018-08-12
 
 - Updates source tag to CentOS 7.5.1804.